# src/docs/api.yaml
openapi: 3.0.0
paths:
...
  /users:
    get:
      description: Az összes felhasználó lekérdezése, csak admin jogosultsággal érhető el
      operationId: findAll
      parameters:
        - name: page
          required: false
          in: query
          description: Page number
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            minimum: 1
            default: 10
            type: number
        - name: username
          required: false
          in: query
          description: keresési feltétel felhasználónév alapján
          schema:
            type: string
        - name: isAdmin
          required: false
          in: query
          description: keresési feltétel isAdmin alapján
          schema:
            type: string
        - name: email
          required: false
          in: query
          description: keresési feltétel email alapján
          schema:
            type: string
      responses:
        '200':
          description: Sikeres lekérdezés, összes felhasználó visszaadva.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
              examples:
                example:
                  summary: Visszatérési példa
                  value:
                    items:
                      - user_id: 37
                        username: Admin
                        email: admin@gmail.com
                        suly: null
                        magassag: null
                        isAdmin: true
                    meta:
                      currentPage: 1
                      itemsPerPage: 1
                      totalItems: 7
                      totalPages: 7
        '401':
          description: Invalid Token
        '403':
          description: Hozzáférés megtagadva - csak Adminok
      security:
        - access-token: []
      summary: Összes felhasználó lekérdezése
      tags:
        - Users
  /users/{id}:
    get:
      description: >-
        Egy felhasználó lekérdezése azonosító alapján. Egy felhasználó csak a
        saját adatait érheti el, míg admin jogosultsággal az összes felhasználó
        adatai lekérdezhetőek.
      operationId: findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Felhasználó adatai visszaadva.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
              examples:
                example:
                  summary: Visszatérési példa
                  value:
                    user_id: 37
                    username: Admin
                    email: admin@gmail.com
                    suly: null
                    magassag: null
                    isAdmin: true
        '401':
          description: Unauthorized
      security:
        - access-token: []
      summary: Egy felhasználó lekérdezése
      tags:
        - Users
    patch:
      description: >-
        Felhasználó adatainak módosítása azonosító alapján. Egy felhasználó csak
        a saját adatait módosíthatja, míg admin jogosultsággal az összes
        felhasználó módosítható.
      operationId: update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: Felhasználó adatai sikeresen módosítva.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
              examples:
                example:
                  summary: Visszatérési példa
                  value:
                    user_id: 37
                    username: Admin
                    email: admin@gmail.com
                    suly: 70
                    magassag: 175
                    isAdmin: true
        '401':
          description: Unauthorized
      security:
        - access-token: []
      summary: Felhasználó adatainak módosítása
      tags:
        - Users
    delete:
      description: >-
        Felhasználó törlése azonosító alapján. Csak admin jogosultsággal érhető
        el.
      operationId: remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Felhasználó sikeresen törölve.
          content:
            application/json:
              schema:
                example:
                  message: User and all related data successfully deleted
        '401':
          description: Unauthorized
        '403':
          description: Hozzáférés megtagadva - csak Adminok
      security:
        - access-token: []
      summary: Felhasználó törlése
      tags:
        - Users
  /users/{id}/bmi:
    get:
      description: >-
        BMI kiszámítása felhasználó azonosító alapján. Egy felhasználó csak a
        saját BMI-jét kérdezheti le, míg admin jogosultsággal az összes
        felhasználó BMI-je lekérdezhető.
      operationId: getBMI
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: BMI kiszámítása sikeres.
          content:
            application/json:
              schema:
                example:
                  bmi: '22.50'
                  type: Normál
        '401':
          description: Unauthorized
      security:
        - access-token: []
      summary: BMI kiszámítása
      tags:
        - Users
  /users/{id}/admin:
    patch:
      description: >-
        Admin jogosultság módosítása felhasználó azonosító alapján. Csak admin
        jogosultsággal érhető el.
      operationId: changeAdmin
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminDto'
      responses:
        '200':
          description: Admin státusz sikeresen megváltoztatva.
          content:
            application/json:
              schema:
                example:
                  message: Admin státusz sikeresen megváltoztatva
        '401':
          description: Unauthorized
        '403':
          description: Hozzáférés megtagadva - csak Adminok
      security:
        - access-token: []
      summary: Admin jogosultság módosítása
      tags:
        - Users
...
  /user-gyakorlat/user/{userId}:
    get:
      description: Lekéri az összes gyakorlatot és azok történetét egy adott felhasználóhoz
      operationId: getUserGyakorlatok
      parameters:
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
        - name: page
          required: false
          in: query
          description: Oldalszám
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Elemek száma oldalanként
          schema:
            minimum: 1
            maximum: 100
            default: 10
            type: number
        - name: isRecord
          required: false
          in: query
          description: ' Ez a tulajdonság arra szolgál, hogy meghatározza, a rekord boolean értékként'
          schema:
            default: false
            type: boolean
        - name: search
          required: false
          in: query
          description: Keresési feltétel gyakorlat neve alapján
          schema:
            type: string
      responses:
        '200':
          description: A gyakorlatok sikeresen lekérve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserGyakorlatokResponseDto'
      security:
        - access-token: []
      summary: Felhasználó összes gyakorlatának lekérése
      tags:
        - User Gyakorlat
...
info:
  title: RepVault
  description: |2-

          RepVault API
      
          A védett végpontok eléréséhez Bearer típusú JWT tokent kell használnia.
          Az authentikáció fül alatt lévő /login végponton keresztül tud tokenhez jutni. Adja meg a felhasználónevét és jelszavát,
          illetve ha még nincs felhasználója, regisztráljon az /register végponton keresztül és aztán jelentkezzen be. 
          Ezt követően kattintson az "Authorize" gombra a Swagger felületén, majd másolja be a tokenjét.
      
          Miután megadta a tokent, a Swagger automatikusan hozzáadja azt az "Authorization" fejléchez, 
          így Ön tesztelheti a védett végpontokat.
        
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:8000
    description: Development Server
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    PaginationMetaDto:
      type: object
      properties:
        currentPage:
          type: number
          example: 1
          description: Aktuális oldalszám
        itemsPerPage:
          type: number
          example: 10
          description: Elemek száma oldalanként
        totalItems:
          type: number
          example: 100
          description: Összes elem száma
        totalPages:
          type: number
          example: 10
          description: Összes oldal száma
        favoriteExercises:
          type: boolean
          example: false
          description: kedvenc edzések
      required:
        - currentPage
        - itemsPerPage
        - totalItems
        - totalPages
...
    User:
      type: object
      properties:
        user_id:
          type: number
          example: 1
          description: A felhasználó azonosítója
        username:
          type: string
          example: teszt
          description: A felhasználó felhasználóneve
        email:
          type: string
          example: teszt@gmail.com
          description: A felhasználó email címe
        password:
          type: string
          example: Teszt1234.
          description: >-
            A felhasználó jelszava (erős jelszó, legalább 8 karakter,
            tartalmazzon kis- és nagybetűt, számot és speciális karaktert)
        isAdmin:
          type: boolean
          example: false
          description: A felhasználó admin jogosultsággal rendelkezik-e
        suly:
          type: number
          example: 80
          description: A felhasználó a súlya
        magassag:
          type: number
          example: 180
          description: A felhasználó a magassága
      required:
        - user_id
        - username
        - email
        - password
        - isAdmin
    UserResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            description: List of users
            type: array
            items:
              $ref: '#/components/schemas/User'
        meta:
          type: object
          description: Pagination metadata
          example:
            currentPage: 1
            itemsPerPage: 1
            totalItems: 7
            totalPages: 7
      required:
        - items
        - meta
    UserDto:
      type: object
      properties:
        user_id:
          type: number
          description: Unique identifier of the user
          example: 37
        username:
          type: string
          description: Username of the user
          example: Admin
        email:
          type: string
          description: Email address of the user
          example: admin@gmail.com
        suly:
          type: number
          description: User weight in kilograms
          example: null
          nullable: true
        magassag:
          type: number
          description: User height in centimeters
          example: null
          nullable: true
        isAdmin:
          type: boolean
          description: Indicates if the user is an administrator
          example: true
      required:
        - user_id
        - username
        - email
        - suly
        - magassag
        - isAdmin
    UpdateUserDto:
      type: object
      properties:
        username:
          type: string
          example: newusername
          description: User name
        suly:
          type: number
          example: 80
          description: User weight
        magassag:
          type: number
          example: 180
          description: User height
        password:
          type: string
          example: newpassword
          description: User password
      required:
        - username
        - suly
        - magassag
        - password
    AdminDto:
      type: object
      properties:
        isAdmin:
          type: boolean
          example: 'true'
          description: IsAdmin?
      required:
        - isAdmin
...
    UserGyakorlat:
      type: object
      properties:
        user_id:
          type: number
          description: A felhasználó azonosítója
          example: 1
        gyakorlat_id:
          type: number
          description: A gyakorlat azonosítója
          example: 1
        personal_best:
          type: number
          description: Személyes rekord súly
          example: 100.5
        last_weight:
          type: number
          description: Utoljára használt súly
          example: 90.5
        last_reps:
          type: number
          description: Utolsó ismétlésszám
          example: 12
        total_sets:
          type: number
          description: Összes elvégzett szett
          example: 50
        gyakorlat:
          type: object
          description: A gyakorlat adatai
        history:
          description: Edzés történet
          type: array
          items:
            type: object
      required:
        - user_id
        - gyakorlat_id
        - total_sets
...
    UserGyakorlatokResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            description: User-Gyakorlat kapcsolatok listája
            type: array
            items:
              $ref: '#/components/schemas/UserGyakorlat'
        meta:
          description: Lapozási metaadatok
          allOf:
            - $ref: '#/components/schemas/PaginationMetaDto'
      required:
        - items
        - meta
...